+++ common/src/main/java/com/faboslav/friendsandfoes/common/init/FriendsAndFoesPointOfInterestTypes.java
@@ -10,7 +10,6 @@ import net.minecraft.core.registries.BuiltInRegistries;
 import net.minecraft.world.entity.ai.village.poi.PoiType;
 import net.minecraft.world.entity.ai.village.poi.PoiTypes;
 import net.minecraft.world.level.block.Block;
-import net.minecraft.world.level.block.ComposterBlock;
 import org.jetbrains.annotations.Nullable;
 
 /**
@@ -47,7 +46,7 @@ public final class FriendsAndFoesPointOfInterestTypes
 	private static Supplier<PoiType> registerPoi(String name, RegistryEntry<Block> block, int maxTickets, int validRange) {
 		return POINT_OF_INTEREST_TYPES.register(name, () -> new PoiType(PoiTypes.getBlockStates(block.get()), maxTickets, validRange));
 	}
-	
+
 	private FriendsAndFoesPointOfInterestTypes() {
 	}
 }