+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 4.0.13
+
+- Fixed illusioner/iceologer mixin crash
+- Fixed lightning rods lightning bolt attraction again (on Fabric)
+- Fixed invalid c tags
+- Fixed missing sound assets warnings
+
 ## 4.0.12
 
 - Fixed raid member mixin crash